chore: merging exporter-console into tracing
