--- packages/opentelemetry-exporter-console/README.md
@@ -1,55 +1,0 @@
-# OpenTelemetry Console Exporter
-[![Gitter chat][gitter-image]][gitter-url]
-[![NPM Published Version][npm-img]][npm-url]
-[![dependencies][dependencies-image]][dependencies-url]
-[![devDependencies][devDependencies-image]][devDependencies-url]
-[![Apache License][license-image]][license-image]
-
-This module provides a helper exporter that can be used when working with Web applications.
-
-## Installation
-
-```bash
-npm install --save @opentelemetry/exporter-console
-```
-
-## Usage
-
-```js
-
-import { ConsoleExporter } from '@opentelemetry/exporter-console';
-import { SimpleSpanProcessor } from '@opentelemetry/tracing';
-import { WebTracer } from '@opentelemetry/web';
-import { DocumentLoad } from '@opentelemetry/plugin-document-load';
-
-const webTracer = new WebTracer({
-  plugins: [
-    new DocumentLoad()
-  ]
-});
-
-webTracer.addSpanProcessor(new SimpleSpanProcessor(new ConsoleExporter()));
-
-// if you run this example you should see the exported spans in console
-
-```
-
-## Useful links
-- For more information on OpenTelemetry, visit: <https://opentelemetry.io/>
-- For more about OpenTelemetry JavaScript: <https://github.com/open-telemetry/opentelemetry-js>
-- For help or feedback on this project, join us on [gitter][gitter-url]
-
-## License
-
-Apache 2.0 - See [LICENSE][license-url] for more information.
-
-[gitter-image]: https://badges.gitter.im/open-telemetry/opentelemetry-js.svg
-[gitter-url]: https://gitter.im/open-telemetry/opentelemetry-node?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge
-[license-url]: https://github.com/open-telemetry/opentelemetry-js/blob/master/LICENSE
-[license-image]: https://img.shields.io/badge/license-Apache_2.0-green.svg?style=flat
-[dependencies-image]: https://david-dm.org/open-telemetry/opentelemetry-js/status.svg?path=packages/opentelemetry-exporter-console
-[dependencies-url]: https://david-dm.org/open-telemetry/opentelemetry-js?path=packages%2Fopentelemetry-exporter-console
-[devDependencies-image]: https://david-dm.org/open-telemetry/opentelemetry-js/dev-status.svg?path=packages/opentelemetry-exporter-console
-[devDependencies-url]: https://david-dm.org/open-telemetry/opentelemetry-js?path=packages%2Fopentelemetry-web&type=dev
-[npm-url]: https://www.npmjs.com/package/@opentelemetry/exporter-console
-[npm-img]: https://badge.fury.io/js/%40opentelemetry%2Fexporter-console.svg
