--- packages/opentelemetry-exporter-console/package.json
@@ -1,72 +1,0 @@
-{
-  "name": "@opentelemetry/exporter-console",
-  "version": "0.1.0",
-  "description": "OpenTelemetry Console Exporter",
-  "main": "build/src/index.js",
-  "types": "build/src/index.d.ts",
-  "repository": "open-telemetry/opentelemetry-js",
-  "scripts": {
-    "test:browser": "nyc karma start --single-run",
-    "tdd": "karma start",
-    "codecov:browser": "nyc report --reporter=json && codecov -f coverage/*.json -p ../../",
-    "clean": "rimraf build/*",
-    "check": "gts check",
-    "compile": "tsc -p .",
-    "watch": "tsc -w",
-    "fix": "gts fix"
-  },
-  "keywords": [
-    "opentelemetry",
-    "web",
-    "tracing",
-    "profiling",
-    "metrics",
-    "stats"
-  ],
-  "author": "OpenTelemetry Authors",
-  "license": "Apache-2.0",
-  "engines": {
-    "node": ">=8.0.0"
-  },
-  "files": [
-    "build/src/**/*.js",
-    "build/src/**/*.d.ts",
-    "doc",
-    "LICENSE",
-    "README.md"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "devDependencies": {
-    "@types/mocha": "^5.2.5",
-    "@types/webpack-env": "1.13.9",
-    "@types/sinon": "^7.0.13",
-    "@babel/core": "^7.6.0",
-    "babel-loader": "^8.0.6",
-    "codecov": "^3.1.0",
-    "gts": "^1.0.0",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-spec-reporter": "^0.0.32",
-    "karma-webpack": "^4.0.2",
-    "mocha": "^6.1.0",
-    "nyc": "^14.1.1",
-    "rimraf": "^3.0.0",
-    "sinon": "^7.5.0",
-    "tslint-consistent-codestyle": "^1.16.0",
-    "tslint-microsoft-contrib": "^6.2.0",
-    "ts-loader": "^6.0.4",
-    "ts-mocha": "^6.0.0",
-    "typescript": "^3.6.3",
-    "webpack": "^4.35.2",
-    "webpack-cli": "^3.3.9",
-    "webpack-merge": "^4.2.2"
-  },
-  "dependencies": {
-    "@opentelemetry/base": "^0.1.0",
-    "@opentelemetry/core": "^0.1.0",
-    "@opentelemetry/tracing": "^0.1.0"
-  }
-}
