--- packages/opentelemetry-exporter-console/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../tsconfig.base",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "build"
-  },
-  "include": [
-    "src/**/*.ts",
-    "test/**/*.ts"
-  ]
-}
