--- packages/opentelemetry-exporter-console/tslint.json
@@ -1,4 +1,0 @@
-{
-  "rulesDirectory": ["node_modules/tslint-microsoft-contrib"],
-  "extends": ["../../tslint.base.js", "./node_modules/tslint-consistent-codestyle"]
-}
